- Add manage.py commands for populate, repopulate, delete entity data - Add EngineConnectionException to prevent silent failing on connection errors - Add None checks and return early to avoid falling into exceptions

+++ app/management/commands/delete_entity_data_datastore.py
@@ -13,7 +13,7 @@ class Command(BaseCommand):
         )
 
     def handle(self, *args, **options):
-        if options['entity_name']:
+        if 'entity_name' in options and options['entity_name']:
             entity_name = options['entity_name']
             db = DataStore()
             db.delete_entity(entity_name=entity_name)
